Rename frontend-queries to parsing-queries
---
