+++ compiler/next/test/frontend/testInteractive.cpp
@@ -17,7 +17,7 @@
  * limitations under the License.
  */
 
-#include "chpl/frontend/frontend-queries.h"
+#include "chpl/frontend/parsing-queries.h"
 #include "chpl/frontend/resolution-queries.h"
 
 #include "chpl/uast/Module.h"
